Separate errors and warnings into two files.
