+++ stitching/stitching_error.py
@@ -1,6 +1,2 @@
 class StitchingError(Exception):
     pass
-
-
-class StitchingWarning(UserWarning):
-    pass
