+++ stitching/stitching_warning.py
@@ -1,0 +1,2 @@
+class StitchingWarning(UserWarning):
+    pass
